Much progress made on QuestHandler.
Quest now boosts foe's battlepoints if the name of the foe matches the
name of the foe on the quest. (test verifies that it works!!)

Made some edits to stage.

Amours changed from object to list to accomodate the Remove() method
(all amours discarded at the end of a quest).

Added some getters for ease of use.